refactor the auto script
	extract the i18n script to common and prepare to the serialized
	job for target building.

branch: dev
user: ysydty
