+++ tools/scripts/common/define.js
@@ -1,7 +1,7 @@
 /**@format */
 
 module.exports = {
-    TIANYU_NATIVE_NATIVE_RES_PATH: "../../native/res",
+    TIANYU_NATIVE_NATIVE_RES_PATH: "../../../native/res",
 
     TIANYU_NATIVE_NATIVE_RES_I18N_NAME: "i18n",
 };
